requirements.txt
elasticsearch
python-docx
PyPDF2
python-pptx
pandas
openpyxl

django
djangorestframework
django-cors-headers
djangorestframework-simplejwt

pytesseract
opencv-python
pillow
numpy

paddlepaddle
paddleocr